usr/lib/hoffui_encoder/requirements.txt
pyinstaller
pillow
ttkbootstrap
tkthread
openai
python-dotenv
git+https://github.com/psf/black
# Note: The above packages are required for the GUI application to function correctly.
# Ensure you have the necessary dependencies installed in your Python environment.
# You can install them using pip:
# pip install -r requirements.txt
